Fixed broken classnames when using BEM with ce_colStart (#14)
* Fixed broken classnames when using BEM with ce_colStart

Using the column element "ce_colStart" with a BEM naming convention results in broken classnames.

Using the classname "ce_colStart--links" for example in combination with "col-6" results in the classname "ce_colStart col-6 --cta". The problem occurs because the class "ce_colStart" is currently being replaced with str_replace.

The fix uses preg_replace to replace "ce_colStart" to prevent this behaviour.

* Fixed broken classnames when using BEM with ce_colStart

Using the column element "ce_colStart" with a BEM naming convention results in broken classnames.

Using the classname "ce_colStart--links" for example in combination with "col-6" results in the classname "ce_colStart col-6 --links". The problem occurs because the class "ce_colStart" is currently being replaced with str_replace.

The fix uses preg_replace to replace "ce_colStart" to prevent this behaviour.